+++ src/main/java/at/porscheinformatik/seleniumcomponents/clarity/ClarityFormControlContainer.java
@@ -18,15 +18,29 @@ import at.porscheinformatik.seleniumcomponents.component.HtmlComponent;
  */
 public class ClarityFormControlContainer extends AbstractSeleniumComponent
 {
+    public static ClarityFormControlContainer within(SeleniumComponent parent)
+    {
+        return new ClarityFormControlContainer(parent, selectByClassName("clr-form-control"));
+    }
 
-    public final HtmlComponent labelComponent =
-        new HtmlComponent(this, WebElementSelector.selectByClassName("clr-control-label"));
+    public static ClarityFormControlContainer byTestId(SeleniumComponent parent, String testId)
+    {
+        return new ClarityFormControlContainer(parent, selectByTestId("clr-form-control", testId));
+    }
+
+    public static ClarityFormControlContainer byText(SeleniumComponent parent, String partialText)
+    {
+        return new ClarityFormControlContainer(parent, selectByText("clr-form-control", partialText));
+    }
+
+    public final HtmlComponent labelComponent = new HtmlComponent(this, selectByClassName("clr-control-label"));
+
     private final SeleniumComponentListFactory<HtmlComponent> validationErrors =
         new SeleniumComponentListFactory<>(this, selectByTagName("clr-control-error"), HtmlComponent::new);
 
     public ClarityFormControlContainer(SeleniumComponent parent)
     {
-        this(parent, WebElementSelector.selectByClassName("clr-form-control"));
+        this(parent, selectByClassName("clr-form-control"));
     }
 
     public ClarityFormControlContainer(SeleniumComponent parent, WebElementSelector selector)
@@ -42,8 +56,6 @@ public class ClarityFormControlContainer extends AbstractSeleniumComponent
     public List<String> getValidationErrors()
     {
         return validationErrors //
-            .findAll()
-            .map(HtmlComponent::getText)
-            .collect(Collectors.toList());
+            .findAll().map(HtmlComponent::getText).collect(Collectors.toList());
     }
 }