+++ src/main/java/at/porscheinformatik/seleniumcomponents/clarity/ClarityRadioContainerComponent.java
@@ -1,7 +1,6 @@
 package at.porscheinformatik.seleniumcomponents.clarity;
 
 import static at.porscheinformatik.seleniumcomponents.SeleniumAsserts.*;
-import static at.porscheinformatik.seleniumcomponents.WebElementSelector.*;
 
 import java.util.Objects;
 import java.util.function.Predicate;
@@ -19,22 +18,36 @@ import at.porscheinformatik.seleniumcomponents.WebElementSelector;
  */
 public class ClarityRadioContainerComponent extends ClarityFormControlContainer
 {
-    private final SeleniumComponentListFactory<ClarityRadioComponent> radios =
-        new SeleniumComponentListFactory<>(this, selectByTagName("clr-radio-wrapper"), ClarityRadioComponent::new);
+    public static ClarityRadioContainerComponent within(SeleniumComponent parent)
+    {
+        return new ClarityRadioContainerComponent(parent, WebElementSelector.selectByTagName("clr-radio-container"));
+    }
 
-    // ---
+    public static ClarityRadioContainerComponent byTestId(SeleniumComponent parent, String testId)
+    {
+        return new ClarityRadioContainerComponent(parent,
+            WebElementSelector.selectByTestId("clr-radio-container", testId));
+    }
+
+    public static ClarityRadioContainerComponent byLabel(SeleniumComponent parent, String partialLabel)
+    {
+        return new ClarityRadioContainerComponent(parent,
+            WebElementSelector.selectByTagNameContainingLabel("clr-radio-container", partialLabel));
+    }
+
+    private final SeleniumComponentListFactory<ClarityRadioComponent> radios =
+        new SeleniumComponentListFactory<>(this, WebElementSelector.selectByTagName("clr-radio-wrapper"),
+            ClarityRadioComponent::new);
 
     public ClarityRadioContainerComponent(SeleniumComponent parent)
     {
-        super(parent, selectByTagName("clr-radio-container"));
+        super(parent, WebElementSelector.selectByTagName("clr-radio-container"));
     }
 
     public ClarityRadioContainerComponent(SeleniumComponent parent, WebElementSelector selector)
     {
         super(parent, selector);
     }
-
-    // ---
 
     public SeleniumComponentListFactory<ClarityRadioComponent> getRadioComponentFactory()
     {
@@ -43,12 +56,55 @@ public class ClarityRadioContainerComponent extends ClarityFormControlContainer
 
     public ClarityRadioComponent getRadioComponentAtIndex(int index)
     {
-        return new ClarityRadioComponent(this, selectByIndex("clr-radio-wrapper", index));
+        return new ClarityRadioComponent(this, WebElementSelector.selectByIndex("clr-radio-wrapper", index));
+    }
+
+    public ClarityRadioComponent getRadioComponentByTestId(String testId)
+    {
+        return ClarityRadioComponent.byTestId(this, testId);
+    }
+
+    public ClarityRadioComponent getRadioComponentByLabel(String label)
+    {
+        return ClarityRadioComponent.byLabel(this, label);
     }
 
     public SeleniumComponentList<ClarityRadioComponent> getRadioComponents()
     {
         return radios.findAll();
+    }
+
+    public void selectByIndex(int index)
+    {
+        ClarityRadioComponent component = getRadioComponentAtIndex(index);
+
+        assertThatSoon(() -> {
+            component.select();
+
+            return component;
+        }, SeleniumMatchers.isSelected());
+    }
+
+    public void selectByTestId(String testId)
+    {
+        ClarityRadioComponent component = getRadioComponentByTestId(testId);
+
+        assertThatSoon(() -> {
+            component.select();
+
+            return component;
+        }, SeleniumMatchers.isSelected());
+    }
+
+    public void selectByLabel(String label)
+    {
+        ClarityRadioComponent component = getRadioComponentByLabel(label);
+
+        assertThatSoon(() -> {
+            component.select();
+
+            return component;
+        }, SeleniumMatchers.isSelected());
     }
 
     public void selectByValue(String value)