+++ src/main/java/at/porscheinformatik/seleniumcomponents/component/RadioComponent.java
@@ -15,6 +15,34 @@ import at.porscheinformatik.seleniumcomponents.WebElementSelector;
  */
 public class RadioComponent extends AbstractSeleniumComponent implements ActiveSeleniumComponent
 {
+    public static CheckboxComponent within(SeleniumComponent parent)
+    {
+        return new CheckboxComponent(parent, selectByCss("input[type=\"radio\"]"));
+    }
+
+    public static CheckboxComponent byTestId(SeleniumComponent parent, String testId)
+    {
+        return new CheckboxComponent(parent, selectByTestId("input[type=\"radio\"]", testId));
+    }
+
+    public static CheckboxComponent byText(SeleniumComponent parent, String partialText)
+    {
+        return new CheckboxComponent(parent, selectByText("input[type=\"radio\"]", partialText));
+    }
+
+    public static CheckboxComponent byLabel(SeleniumComponent parent, String partialText)
+    {
+        String xpath = String.format(".//label[contains(., '%s')]/input[@type='radio']", partialText);
+
+        return new CheckboxComponent(parent, selectByXPath(xpath));
+    }
+
+    public static CheckboxComponent byValue(SeleniumComponent parent, String value)
+    {
+        String xpath = String.format(".//input[@type='checkbox' and @value='%s']", value);
+
+        return new CheckboxComponent(parent, selectByXPath(xpath));
+    }
 
     /**
      * Creates the component.